requirements.txt
absl-py==1.3.0
anyio==3.6.1
APScheduler==3.6.3
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
bst==0.5.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Django==4.0.4
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
fanstatic==1.2
fastapi==0.85.1
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==22.9.24
fonttools==4.38.0
gast==0.4.0
gdown==4.5.3
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
js==1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lightgbm==3.3.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mediapipe==0.8.11
mtcnn==0.1.1
numpy==1.23.3
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==3.19.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
selenium==4.4.3
shutilwhich==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.20.4
telegram==0.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.18.3
WebOb==1.8.7
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.9.0
